MWE of push_metadata_through
