+++ pyxem/utils/__init__.py
@@ -15,3 +15,34 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
+
+def push_metadata_through(dummy,*args,**kwargs):
+    """
+    This function pushes loaded metadata through to pyxem objects, it is to be used for one
+    purpose, see the __init__ of ElectronDiffraction for an example.
+
+    Parameters
+    ----------
+    dummy :
+        This will always be the "self" of the object to be initialised
+
+    args : list
+        This will always be the "args" of the object to be initialised
+
+    kwargs : dict
+        This will always be the "args" of the object to be initialised
+
+    Returns
+    -------
+    dummy,args,kwargs :
+        The input variables, adjusted correctly
+    """
+    try:
+        meta_dict = args[0].metadata.as_dictionary()
+        kwargs.update({'metadata': meta_dict})
+    except AttributeError:
+        pass  # this is because a numpy array has been passed
+    except IndexError:
+        pass  # this means that map continues to work.
+
+    return dummy,args,kwargs
